Address PR #202 second review: KeyError guard, loaders module, CI, perf, tests
Bugs:
- settings: guard _apply_model_path_override against unknown default_mbert_version (KeyError) (#1)
- run_eval: warn when scoring the v2.5 baseline default instead of production weights (#3)

Quality:
- batching_service: log a warning when EnhancedPreprocessor import fails (no silent disable) (#4)
- extract shared dataset loaders to evals/loaders.py; run_eval + calibrate import it normally (drops importlib hack) (#5)
- drop unreferenced intermediate summaries summary_v2.5_mishra/v2.6/v2.6_big.json (#6)
- health: de-rot the architecture/tokenizer-key comment (no review-number ref) (#9)
- enhanced_preprocessing: INVISIBLE_CHARS as \u escapes with codepoint comments (set unchanged) (minor)

Performance:
- batching_service: ASCII fast-path in _normalize_text skips normalisation for typical traffic (#7)

API:
- response_models: mark HealthResponse.version Field(deprecated=True) (#8)

Tests/CI:
- add GitHub Actions unit-test workflow running pytest (#12)
- add settings-override tests incl. KeyError guard + per-instance isolation (#13)
- add ASCII fast-path test; keep raise-swallow test on non-ASCII input (#7)

Docs:
- TIER1.md: loosen torch pin to stable; flag in-domain validation caveat (minor)
- dataset/README_SYNTHETIC.md: mark synthetic data, warn against pipeline misuse (#11)
