+++ .github/workflows/tests.yml
@@ -1,0 +1,39 @@
+name: Unit Tests
+
+# Runs the fast, offline unit tests (no model weights / LFS needed). The
+# batching + normalisation tests use stubs and stdlib; settings tests construct
+# the config object directly.
+on:
+  push:
+    branches: ["**"]
+    paths:
+      - "src/api_interface/**"
+      - ".github/workflows/tests.yml"
+  pull_request:
+    paths:
+      - "src/api_interface/**"
+      - ".github/workflows/tests.yml"
+
+jobs:
+  unit-tests:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          lfs: false  # tests stub the model; weights not required
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+          cache: pip
+
+      - name: Install test dependencies
+        run: |
+          python -m pip install --upgrade pip
+          # Targeted deps for the unit suite (CPU torch wheel on Linux).
+          pip install "torch>=2.3" "transformers>=4.40" \
+                      "pydantic>=2.7" pydantic-settings \
+                      pytest pytest-asyncio
+
+      - name: Run unit tests
+        run: pytest src/api_interface/tests/ -v
