--- evals/results/summary_v2.5_mishra.json
@@ -1,49 +1,0 @@
-{
-  "mishra": {
-    "n": 5971,
-    "accuracy_3class": 0.8881,
-    "accuracy_binary_block": 0.9915,
-    "per_class": {
-      "ham": {
-        "precision": 0.9983,
-        "recall": 0.9911,
-        "f1": 0.9947,
-        "support": 4844
-      },
-      "spam": {
-        "precision": 0.4236,
-        "recall": 0.9918,
-        "f1": 0.5936,
-        "support": 489
-      },
-      "phishing": {
-        "precision": 1.0,
-        "recall": 0.0266,
-        "f1": 0.0519,
-        "support": 638
-      }
-    },
-    "confusion": {
-      "ham->ham": 4801,
-      "ham->spam": 43,
-      "phishing->ham": 4,
-      "phishing->phishing": 17,
-      "phishing->spam": 617,
-      "spam->ham": 4,
-      "spam->spam": 485
-    },
-    "by_category": {
-      "mishra": {
-        "acc": 0.8881,
-        "n": 5971
-      }
-    },
-    "by_language": {
-      "en": {
-        "acc": 0.8881,
-        "n": 5971
-      }
-    },
-    "inference_seconds": 212.5
-  }
-}
