--- evals/results/summary_v2.6_big.json
@@ -1,397 +1,0 @@
-{
-  "uci": {
-    "n": 5574,
-    "accuracy_3class": 0.9923,
-    "accuracy_binary_block": 0.9923,
-    "per_class": {
-      "ham": {
-        "precision": 0.9998,
-        "recall": 0.9913,
-        "f1": 0.9955,
-        "support": 4827
-      },
-      "spam": {
-        "precision": 0.9626,
-        "recall": 0.9987,
-        "f1": 0.9803,
-        "support": 747
-      },
-      "phishing": {
-        "precision": 0.0,
-        "recall": 0.0,
-        "f1": 0.0,
-        "support": 0
-      }
-    },
-    "confusion": {
-      "ham->ham": 4785,
-      "ham->phishing": 13,
-      "ham->spam": 29,
-      "spam->ham": 1,
-      "spam->spam": 746
-    },
-    "by_category": {
-      "uci": {
-        "acc": 0.9923,
-        "n": 5574
-      }
-    },
-    "by_language": {
-      "en": {
-        "acc": 0.9923,
-        "n": 5574
-      }
-    },
-    "inference_seconds": 97.6
-  },
-  "imc25": {
-    "n": 8007,
-    "accuracy_3class": 0.4193,
-    "accuracy_binary_block": 0.7818,
-    "per_class": {
-      "ham": {
-        "precision": 0.0,
-        "recall": 0.0,
-        "f1": 0.0,
-        "support": 0
-      },
-      "spam": {
-        "precision": 0.0971,
-        "recall": 0.8042,
-        "f1": 0.1733,
-        "support": 383
-      },
-      "phishing": {
-        "precision": 0.9871,
-        "recall": 0.3999,
-        "f1": 0.5692,
-        "support": 7624
-      }
-    },
-    "confusion": {
-      "phishing->ham": 1712,
-      "phishing->phishing": 3049,
-      "phishing->spam": 2863,
-      "spam->ham": 35,
-      "spam->phishing": 40,
-      "spam->spam": 308
-    },
-    "by_category": {
-      "banking": {
-        "acc": 0.4172,
-        "n": 3607
-      },
-      "delivery": {
-        "acc": 0.4753,
-        "n": 930
-      },
-      "government": {
-        "acc": 0.4803,
-        "n": 762
-      },
-      "hey mum/dad": {
-        "acc": 0.541,
-        "n": 61
-      },
-      "others": {
-        "acc": 0.3182,
-        "n": 1606
-      },
-      "spam": {
-        "acc": 0.8042,
-        "n": 383
-      },
-      "telecom": {
-        "acc": 0.3303,
-        "n": 551
-      },
-      "unknown": {
-        "acc": 0.3333,
-        "n": 21
-      },
-      "wrong number": {
-        "acc": 0.0349,
-        "n": 86
-      }
-    },
-    "by_language": {
-      "Afrikaans": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Albanian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Arabic": {
-        "acc": 0.2857,
-        "n": 14
-      },
-      "Bengali": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Bulgarian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Catalan": {
-        "acc": 0.2857,
-        "n": 7
-      },
-      "Chinese": {
-        "acc": 0.3636,
-        "n": 11
-      },
-      "Croatian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Czech": {
-        "acc": 0.4706,
-        "n": 17
-      },
-      "Danish": {
-        "acc": 0.4615,
-        "n": 26
-      },
-      "Divehi": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Dutch": {
-        "acc": 0.5381,
-        "n": 459
-      },
-      "English": {
-        "acc": 0.4157,
-        "n": 5215
-      },
-      "Estonian": {
-        "acc": 1.0,
-        "n": 2
-      },
-      "Filipino": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Finnish": {
-        "acc": 0.0,
-        "n": 3
-      },
-      "French": {
-        "acc": 0.3418,
-        "n": 275
-      },
-      "Ganda": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "German": {
-        "acc": 0.3089,
-        "n": 191
-      },
-      "Greek": {
-        "acc": 0.05,
-        "n": 20
-      },
-      "Gujarati": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Hebrew": {
-        "acc": 0.0,
-        "n": 2
-      },
-      "Hindi": {
-        "acc": 0.3659,
-        "n": 41
-      },
-      "Hungarian": {
-        "acc": 0.3333,
-        "n": 6
-      },
-      "Icelandic": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Igbo": {
-        "acc": 0.0,
-        "n": 2
-      },
-      "Indonesian": {
-        "acc": 0.1951,
-        "n": 82
-      },
-      "Italian": {
-        "acc": 0.462,
-        "n": 158
-      },
-      "Japanese": {
-        "acc": 0.5082,
-        "n": 61
-      },
-      "Javanese": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Kannada": {
-        "acc": 0.0,
-        "n": 2
-      },
-      "Kinyarwanda": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Korean": {
-        "acc": 0.2857,
-        "n": 7
-      },
-      "Latvian": {
-        "acc": 0.0,
-        "n": 2
-      },
-      "Lithuanian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Malay": {
-        "acc": 0.1389,
-        "n": 36
-      },
-      "Maltese": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Marathi": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Mixed (Dutch/French)": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Mongolian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Norwegian": {
-        "acc": 0.5,
-        "n": 8
-      },
-      "Persian": {
-        "acc": 0.0,
-        "n": 2
-      },
-      "Polish": {
-        "acc": 0.3548,
-        "n": 31
-      },
-      "Portuguese": {
-        "acc": 0.3788,
-        "n": 66
-      },
-      "Punjabi": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Romanian": {
-        "acc": 0.25,
-        "n": 8
-      },
-      "Russian": {
-        "acc": 0.5,
-        "n": 2
-      },
-      "Serbian": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Sinhala": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Slovak": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Slovenian": {
-        "acc": 1.0,
-        "n": 3
-      },
-      "Spanish": {
-        "acc": 0.4836,
-        "n": 1096
-      },
-      "Swahili": {
-        "acc": 0.2857,
-        "n": 7
-      },
-      "Swedish": {
-        "acc": 0.2308,
-        "n": 13
-      },
-      "Tagalog": {
-        "acc": 0.3158,
-        "n": 19
-      },
-      "Tamil": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Telugu": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Thai": {
-        "acc": 0.1333,
-        "n": 15
-      },
-      "Turkish": {
-        "acc": 0.2273,
-        "n": 22
-      },
-      "Ukrainian": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Unknown": {
-        "acc": 0.5,
-        "n": 4
-      },
-      "Urdu": {
-        "acc": 0.1111,
-        "n": 18
-      },
-      "Uzbek": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Vietnamese": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Welsh": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "Yoruba": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "Zulu": {
-        "acc": 0.0,
-        "n": 1
-      },
-      "nl_fr": {
-        "acc": 1.0,
-        "n": 1
-      },
-      "unknown": {
-        "acc": 0.2917,
-        "n": 24
-      }
-    },
-    "inference_seconds": 227.0
-  }
-}
